@@ -1,11 +1,16 @@
-import { SurahConstant } from './constant/surah'
-const locale = require('./locale/id')
+/* eslint-disable no-unused-vars */
+import NuxtConfiguration from '@nuxt/config'
 
-const pkg = require('./package')
 const path = require('path')
 
+const SurahConstant = require('./constant/surah')
+const locale = require('./locale/id')
+const pkg = require('./package')
+
+const PROD_PATH = 'https://www.quran-offline.xyz/'
+
 const getOfflineAssets = () => {
-  let res = [
+  const res = [
     '/favicon.ico',
     '/favicon-16x16.png',
     '/favicon-32x32.png',
@@ -23,7 +28,7 @@ const getOfflineAssets = () => {
 }
 
 const routes = () => {
-  let res = [
+  const res = [
     '/',
     '/about',
     '/all-surah',
@@ -37,7 +42,7 @@ const routes = () => {
   ]
   for (let i = 1; i < 115; i++) {
     res.push(`/${i}`)
-    const surahObj = SurahConstant.surahArray.find(item => item.index === i)
+    const surahObj = SurahConstant.find(item => item.index === i)
     if (surahObj) {
       for (let j = 1; j < surahObj.ayah_count + 1; j++) {
         // res.push(`/${i}/${j}`)
@@ -48,20 +53,22 @@ const routes = () => {
 }
 
 const routesSitemap = () => {
-  let res = []
-  routes().forEach(el => {
-    const item = {}
-    item.url = el
-    item.changefreq = 'daily'
-    item.priority = 1
-    item.lastmodISO = String(new Date().toISOString())
+  const res = []
+  routes().forEach((el) => {
+    const item = {
+      url: el,
+      changefreq: 'daily',
+      priority: 1,
+      lastmodISO: String(new Date().toISOString())
+    }
     res.push(item)
   })
   return res
 }
 
-module.exports = {
-  mode: 'spa',
+const config: NuxtConfiguration = {
+  debug: true,
+  mode: 'universal',
   /*
    ** Headers of the page
    */
@@ -76,13 +83,13 @@ module.exports = {
       { hid: 'og:image', property: 'og:image', content: '/icon.png' },
       { hid: 'og:title', property: 'og:title', content: 'Quran Offline' },
       { hid: 'og:description', property: 'og:description', content: pkg.description },
-      { hid: 'og:url', property: 'og:url', content: 'https://quran-offline.netlify.com/' },
+      { hid: 'og:url', property: 'og:url', content: PROD_PATH },
 
       { hid: 'twitter:card', name: 'twitter:card', content: 'summary_large_image' },
       { hid: 'twitter:image:src', name: 'twitter:image:src', content: '/icon.png' },
       { hid: 'twitter:title', name: 'twitter:title', content: 'Quran Offline' },
       { hid: 'twitter:description', name: 'twitter:description', content: pkg.description },
-      { hid: 'twitter:url', name: 'twitter:url', content: 'https://quran-offline.netlify.com/' },
+      { hid: 'twitter:url', name: 'twitter:url', content: PROD_PATH },
 
       { hid: 'google-site-verification', name: 'google-site-verification', content: 'jW7EK0wGpuReuZkQ-q900J7Z0KbCD9CCAZybfwcPe_U' }
     ],
@@ -129,23 +136,7 @@ module.exports = {
    */
   modules: [
     '@nuxtjs/pwa',
-    '@nuxtjs/sitemap',
-    ['nuxt-i18n', {
-      baseUrl: 'https://quran-offline.netlify.com/',
-      locales: [
-        {
-          code: 'id',
-          iso: 'id-ID'
-        }
-      ],
-      defaultLocale: 'id',
-      vueI18n: {
-        fallbackLocale: 'id',
-        messages: {
-          id: locale
-        }
-      }
-    }]
+    '@nuxtjs/sitemap'
   ],
 
   sitemap: {
@@ -178,9 +169,9 @@ module.exports = {
     /*
      ** You can extend webpack config here
      */
-    extend (config, ctx) {
+    extend(config, ctx) {
       // Run ESLint on save
-      config.resolve.alias['icons'] = path.resolve(__dirname, 'node_modules/vue-ionicons/dist')
+      config.resolve.alias.icons = path.resolve(__dirname, 'node_modules/vue-ionicons/dist')
       if (ctx.isDev && ctx.isClient) {
         config.module.rules.push({
           enforce: 'pre',
@@ -192,3 +183,5 @@ module.exports = {
     }
   }
 }
+
+export default config
