@@ -1,19 +1,15 @@
 module.exports = {
-  root: true,
-  env: {
-    browser: true,
-    node: true
-  },
+  plugins: ['@typescript-eslint'],
   parserOptions: {
-    parser: 'babel-eslint'
+    parser: '@typescript-eslint/parser'
   },
-  extends: ['standard', 'plugin:vue/recommended'],
-  // required to lint *.vue files
-  plugins: ['vue'],
-  // add your custom rules here
+  extends: [
+    '@nuxtjs'
+  ],
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-unused-vars': 'off',
     'vue/component-name-in-template-casing': [
       'error',
       'PascalCase',
@@ -29,4 +25,4 @@ module.exports = {
       }
     ]
   }
-};
+}
