@@ -1,7 +1,5 @@
 <template>
-  <header
-    id="header"
-    class="header">
+  <header id="header" class="header">
     <div class="header__sticky">
       <div class="header__nav">
         <a
@@ -10,9 +8,7 @@
           href="javascript:void(0)"
           title="Open Sidebar"
           @click="toggleSidebar">
-          <MdMenuIcon
-            w="30px"
-            h="30px" />
+          <MdMenuIcon w="30px" h="30px" />
         </a>
         <a
           v-else
@@ -20,60 +16,39 @@
           class="header__hamburger"
           title="Kembali"
           @click="backToPreviousPage">
-          <MdArrowBackIcon
-            w="30px"
-            h="30px" />
+          <MdArrowBackIcon w="30px" h="30px" />
         </a>
       </div>
-      <div
-        class="header__content">
-        <nuxt-link
-          v-if="isHomePage"
-          to="/"
-          class="header__title">
+      <div class="header__content">
+        <nuxt-link v-if="isHomePage" to="/" class="header__title">
           <h1>{{ headerTitle }}</h1>
         </nuxt-link>
         <h1 v-else>
           {{ headerTitle }}
         </h1>
       </div>
-      <div
-        v-if="!isHomePage"
-        class="header__nav pointer"
-        @click="toggleMenuRight">
-        <MdMoreIcon
-          w="30px"
-          h="30px" />
+      <div v-if="!isHomePage" class="header__nav pointer" @click="toggleMenuRight">
+        <MdMoreIcon w="30px" h="30px" />
       </div>
-      <div
-        v-else
-        class="header__nav">
-        &nbsp;
+      <div v-else class="header__nav">
+&nbsp;
       </div>
     </div>
-    <nav
-      v-show="isShowMenu"
-      class="menu_right">
+    <nav v-show="isShowMenu" class="menu_right">
       <ul>
         <li>
-          <div
-            class="menu_link"
-            @click="navigateTo('/')">
-            {{ $t('home') }}
+          <div class="menu_link" @click="navigateTo('/')">
+            Beranda
           </div>
         </li>
         <li>
-          <div
-            class="menu_link"
-            @click="navigateTo('/settings')">
-            {{ $t('setting') }}
+          <div class="menu_link" @click="navigateTo('/settings')">
+            Setelan
           </div>
         </li>
         <li>
-          <div
-            class="menu_link"
-            @click="navigateTo('/about')">
-            {{ $t('about') }}
+          <div class="menu_link" @click="navigateTo('/about')">
+            Tentang
           </div>
         </li>
       </ul>
@@ -81,8 +56,9 @@
   </header>
 </template>
 
-<script>
-import { mapState, mapMutations } from 'vuex'
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator'
+import { State, Mutation } from 'vuex-class'
 
 import MdMenuIcon from 'vue-ionicons/dist/js/md-menu'
 import MdArrowBackIcon from 'vue-ionicons/dist/js/md-arrow-back'
@@ -90,61 +66,55 @@ import MdMoreIcon from 'vue-ionicons/dist/js/md-more'
 
 import { AppConstant } from '../constant/index'
 
-export default {
-  name: 'BaseHeader',
+@Component({
   components: {
     MdMenuIcon,
     MdArrowBackIcon,
     MdMoreIcon
-  },
-  data () {
-    return {
-      isShowMenu: false
+  }
+})
+
+export default class BaseHeader extends Vue {
+  isShowMenu = false;
+
+  @State headerTitle;
+  @Mutation setShowSidebar;
+
+  get isHomePage(): boolean {
+    return this.headerTitle === AppConstant.TITLE
+  }
+
+  navigateTo(link): void {
+    this.toggleMenuRight()
+    if (link.indexOf('http') >= 0) {
+      window.location.href = link
+    } else {
+      this.$router.push(link)
     }
-  },
-  computed: {
-    ...mapState([
-      'headerTitle'
-    ]),
-    isHomePage () {
-      return this.headerTitle === AppConstant.TITLE
+  }
+
+  toggleMenuRight(): void {
+    this.isShowMenu = !this.isShowMenu
+    if (this.isShowMenu) {
+      setTimeout(() => {
+        this.isShowMenu = false
+      }, 2000)
     }
-  },
-  methods: {
-    ...mapMutations([
-      'setShowSidebar'
-    ]),
-    navigateTo (link) {
-      this.toggleMenuRight()
-      if (link.indexOf('http') >= 0) {
-        window.location.href = link
-      } else {
-        this.$router.push(link)
-      }
-    },
-    toggleMenuRight () {
-      this.isShowMenu = !this.isShowMenu
-      if (this.isShowMenu) {
-        setTimeout(() => {
-          this.isShowMenu = false
-        }, 2000)
-      }
-    },
-    toggleSidebar () {
-      this.setShowSidebar(true)
-    },
-    backToPreviousPage () {
-      window.history.length > 1
-        ? this.$router.go(-1)
-        : this.$router.push('/')
-    }
+  }
+
+  toggleSidebar() {
+    this.setShowSidebar(true)
+  }
+
+  backToPreviousPage() {
+    window.history.length > 1 ? this.$router.go(-1) : this.$router.push('/')
   }
 }
 </script>
 
 <style lang="scss" scoped>
-@import '../assets/header.scss';
-.menu_right{
+@import "../assets/header.scss";
+.menu_right {
   position: fixed;
   top: 50px;
   right: 30px;
@@ -155,9 +125,9 @@ export default {
     padding: 0;
     list-style: none;
 
-    li{
+    li {
       border-bottom: 1px solid #41b883;
-      .menu_link{
+      .menu_link {
         display: block;
         background-color: #1a1a1a;
         color: $theme;
@@ -167,7 +137,7 @@ export default {
         display: flex;
         align-items: center;
       }
-      &:last-child{
+      &:last-child {
         border-bottom: 0;
       }
     }
