@@ -10,21 +10,12 @@
           {{ Number(index) }}
         </div>
         <div class="verse__header--right">
-          <div
-            class="verse__header_icon"
-            @click="onClickAudioIcon(Number(index))">
-            <MdVolumeHighIcon
-              w="2em"
-              h="2em" />
+          <div class="verse__header_icon" @click="onClickAudioIcon(Number(index))">
+            <MdVolumeHighIcon w="2em" h="2em" />
 
-            <div
-              class="tooltip"
-              :class="{'show' : Number(clickedAudioIcon) === Number(index)}">
-              <ul
-                v-for="audio in audios"
-                :key="audio.id">
-                <li
-                  @click="onClickAudioItem(audio, Number(index))">
+            <div class="tooltip" :class="{'show' : Number(clickedAudioIcon) === Number(index)}">
+              <ul v-for="audio in audios" :key="audio.id">
+                <li @click="onClickAudioItem(audio, Number(index))">
                   {{ audio.text }}
                 </li>
               </ul>
@@ -34,31 +25,23 @@
           <div
             class="verse__header_icon"
             @click="doSetLastReadVerse({ surah: surahId, verse: Number(index) })">
-            <MdBookmarkIcon
-              w="2em"
-              h="2em" />
+            <MdBookmarkIcon w="2em" h="2em" />
           </div>
 
           <div
             v-if="isSupportWebShare"
             class="verse__header_icon"
             @click="shareVerse(verse, Number(index))">
-            <MdShareIcon
-              w="2em"
-              h="2em" />
+            <MdShareIcon w="2em" h="2em" />
           </div>
         </div>
       </div>
       <div class="divider clearfix">
-        <div
-          class="verse__arabic font-arabic"
-          dir="rtl">
+        <div class="verse__arabic font-arabic" dir="rtl">
           {{ verse }}
         </div>
       </div>
-      <div
-        v-if="settingShowTranslation"
-        class="divider clearfix">
+      <div v-if="settingShowTranslation" class="divider clearfix">
         <div class="verse__bold">
           <b>Terjemah:</b>
         </div>
@@ -66,9 +49,7 @@
           {{ getTranslation(index) }}
         </div>
       </div>
-      <div
-        v-if="settingShowTafsir"
-        class="divider clearfix">
+      <div v-if="settingShowTafsir" class="divider clearfix">
         <div class="verse__bold">
           <b>Tafsir dari Kemenag:</b>
         </div>
@@ -83,8 +64,10 @@
   </div>
 </template>
 
-<script>
-import { mapActions, mapState } from 'vuex'
+<script lang="ts">
+import { Component, Prop, Watch, Vue } from 'vue-property-decorator'
+import { State, Action } from 'vuex-class'
+
 import MdShareIcon from 'vue-ionicons/dist/js/md-share'
 import MdBookmarkIcon from 'vue-ionicons/dist/js/md-bookmark'
 import MdVolumeHighIcon from 'vue-ionicons/dist/js/md-volume-high'
@@ -92,89 +75,78 @@ import MurotalConstant from '../constant/murotal'
 
 const randomVerse = 1000000
 
-export default {
-  name: 'VerseCard',
+@Component({
   components: {
     MdShareIcon,
     MdBookmarkIcon,
     MdVolumeHighIcon
-  },
-  props: {
-    verseArray: {
-      type: [Object, Array],
-      default: () => ({})
-    },
-    translations: {
-      type: Object,
-      default: () => ({})
-    },
-    tafsir: {
-      type: Object,
-      default: () => ({})
-    },
-    surahId: {
-      type: Number,
-      default: 0
-    }
-  },
-  data () {
-    return {
-      audios: MurotalConstant.availableAudio,
-      clickedAudioIcon: randomVerse,
-      timeout: null
-    }
-  },
-  computed: {
-    ...mapState([
-      'surahFavorite',
-      'isSupportWebShare',
-      'settingShowTranslation',
-      'settingShowTafsir'
-    ])
-  },
-  methods: {
-    ...mapActions([
-      'setLastReadVerse',
-      'showNotification',
-      'shareViaWebshare'
-    ]),
-    onClickAudioIcon (verse) {
-      if (this.timeout) clearTimeout(this.timeout)
+  }
+})
+export default class VerseCard extends Vue {
+  audios = MurotalConstant.availableAudio;
+  clickedAudioIcon = randomVerse;
+  timeout;
 
-      this.clickedAudioIcon = verse
-      this.timeout = setTimeout(() => {
-        this.clickedAudioIcon = randomVerse
-      }, 5000)
-    },
-    onClickAudioItem (audioObj, verse) {
-      if (audioObj.id === 'kemenag') {
-        const hrefAudio = MurotalConstant.getAudioFromKemenag(this.surahId, verse)
-        window.location.href = hrefAudio
-      }
-    },
-    getTranslation (indexVerse) {
-      return this.translations.id.text[indexVerse]
-    },
-    getTafsir (indexVerse) {
-      return this.tafsir.id.kemenag.text[indexVerse]
-    },
-    doSetLastReadVerse (data) {
-      this.setLastReadVerse(data)
-      this.showNotification({
-        title: 'Pesan Sukses',
-        message: 'Ayat telah ditambahkan ke terakhir dibaca.'
-      })
-    },
-    shareVerse (verse, index) {
-      const data = {
-        title: `QS ${this.surahId}:${index}`,
-        text: `${verse}
+  @Prop({ type: [Object, Array], default: () => ({}) }) readonly verseArray!:
+    | any
+    | any[];
+  @Prop({ type: Object, default: () => ({}) }) readonly translations!: any;
+  @Prop({ type: Object, default: () => ({}) }) readonly tafsir!: any;
+  @Prop({ type: Number, default: 0 }) readonly surahId!: number;
+
+  @State surahFavorite;
+  @State isSupportWebShare;
+  @State settingShowTranslation;
+  @State settingShowTafsir;
+
+  @Action setLastReadVerse;
+  @Action showNotification;
+  @Action shareViaWebshare;
+
+  onClickAudioIcon(verse) {
+    if (this.timeout) clearTimeout(this.timeout)
+
+    this.clickedAudioIcon = verse
+    this.timeout = setTimeout(() => {
+      this.clickedAudioIcon = randomVerse
+    }, 5000)
+  }
+
+  onClickAudioItem(audioObj, verse) {
+    if (audioObj.id === 'kemenag') {
+      const hrefAudio = MurotalConstant.getAudioFromKemenag(
+        this.surahId,
+        verse
+      )
+      window.location.href = hrefAudio
+    }
+  }
+
+  getTranslation(indexVerse) {
+    return this.translations.id.text[indexVerse]
+  }
+
+  getTafsir(indexVerse) {
+    return this.tafsir.id.kemenag.text[indexVerse]
+  }
+
+  doSetLastReadVerse(data) {
+    this.setLastReadVerse(data)
+    this.showNotification({
+      title: 'Pesan Sukses',
+      message: 'Ayat telah ditambahkan ke terakhir dibaca.'
+    })
+  }
+
+  shareVerse(verse, index) {
+    const data = {
+      title: `QS ${this.surahId}:${index}`,
+      text: `${verse}
 
         ${this.getTranslation(index)} (QS ${this.surahId}:${index})`,
-        url: `https://quran-offline.netlify.com/${this.surahId}#verse-${index}`
-      }
-      this.shareViaWebshare(data)
+      url: `https://quran-offline.netlify.com/${this.surahId}#verse-${index}`
     }
+    this.shareViaWebshare(data)
   }
 }
 </script>
@@ -183,7 +155,7 @@ export default {
 .verse {
   text-align: left;
 
-  &__header{
+  &__header {
     display: flex;
     align-items: center;
     justify-content: space-between;
@@ -193,7 +165,7 @@ export default {
       justify-content: space-between;
     }
     &_icon {
-      padding: 0 .5em;
+      padding: 0 0.5em;
       position: relative;
     }
   }
@@ -202,7 +174,7 @@ export default {
     width: 100%;
     float: right;
     text-align: right;
-    margin-top: .25em;
+    margin-top: 0.25em;
   }
   &__translation {
     font-size: 1rem;
@@ -215,18 +187,18 @@ export default {
     width: 100%;
     margin-top: 1.5em;
   }
-  &__bold{
+  &__bold {
     font-weight: bold;
     margin-top: 1.5em;
     line-height: 1.5;
   }
-  &__small{
+  &__small {
     font-size: 0.8rem;
     margin-top: 1.5em;
     font-style: italic;
   }
 }
-.tooltip{
+.tooltip {
   position: absolute;
   opacity: 0;
   visibility: hidden;
@@ -248,7 +220,7 @@ export default {
     li {
       padding: 15px 25px;
       border-bottom: 1px solid #41b883;
-      &:last-child{
+      &:last-child {
         border-bottom: 0;
       }
     }
@@ -258,7 +230,7 @@ export default {
   visibility: visible;
   opacity: 1;
 }
-.tooltip{
+.tooltip {
   position: absolute;
   opacity: 0;
   visibility: hidden;
@@ -280,7 +252,7 @@ export default {
     li {
       padding: 15px 25px;
       border-bottom: 1px solid #41b883;
-      &:last-child{
+      &:last-child {
         border-bottom: 0;
       }
     }
