@@ -1,7 +1,11 @@
+const isProd = (process.env.NODE_ENV === 'production')
+const PROD_PATH = 'https://www.quran-offline.xyz/'
+const rootPath = isProd ? PROD_PATH : PROD_PATH
+
 export default {
-  SURAH_INFO: '/data/surah-info.json',
-  SURAH_BY_ID: (id) => `/data/surah/${id}.json`,
-  AYAT_KURSI: '/data/ayat-kursi.json',
-  ASMAUL_HUSNA: '/data/asmaul-husna.json',
-  DAILY_DOA: '/data/daily-doa.json'
+  SURAH_INFO: `${rootPath}data/surah-info.json`,
+  SURAH_BY_ID: id => `${rootPath}data/surah/${id}.json`,
+  AYAT_KURSI: `${rootPath}data/ayat-kursi.json`,
+  ASMAUL_HUSNA: `${rootPath}data/asmaul-husna.json`,
+  DAILY_DOA: `${rootPath}data/daily-doa.json`
 }
