@@ -18,7 +18,7 @@ module.exports = {
     about: `Tentang | ${title}`,
     home: `Baca Al-Qur'an dimana saja, langsung dari web browser Anda | ${title}`,
     recommendation: `Baca surat rekomendasi dalam Al-Qur'an | ${title}`,
-    favorite: `Baca surat favorit dalam Al-Qur'an | ${title}`,
+    favorite: `Baca berbagai surat favorit dalam Al-Qur'an | ${title}`,
     dailyDoa: `Daftar bacaan do'a sehari-hari beserta terjemahan | ${title}`,
     ayatKursi: `Bacaan dan terjemah ayat kursi | ${title}`,
     asmaulHusna: `Daftar lengkap asmaul husna beserta terjemahan | ${title}`,
