@@ -23,20 +23,21 @@
 </template>
 
 <script>
-import { mapActions, mapMutations, mapState } from 'vuex'
+import { mapState } from 'vuex'
 
 import SurahCard from '../components/SurahCard.vue'
 import { __isNotEmptyString, __normalizeText } from '../utils/index'
+import { getAllSurah } from '../services/index'
 
 export default {
   name: 'PageAllSurah',
-  head () {
+  head() {
     return this.metaHead
   },
   components: {
     SurahCard
   },
-  data () {
+  data() {
     return {
       loading: true,
       searchText: ''
@@ -44,11 +45,10 @@ export default {
   },
   computed: {
     ...mapState([
-      'settingActiveTheme',
-      'allSurahList'
+      'settingActiveTheme'
     ]),
-    metaHead () {
-      const title = this.$t('pageTitle.allSurah')
+    metaHead() {
+      const title = 'Daftar semua surat dalam Al-Qur\'an | Qur\'an Offline'
       return {
         title,
         meta: [
@@ -58,13 +58,13 @@ export default {
         ]
       }
     },
-    filteredSurah () {
+    filteredSurah() {
       if (__isNotEmptyString(this.searchText) && this.searchText.length >= 3) {
-        return this.allSurahList.filter(item => {
-          let predicateTranslation = __normalizeText(item.translation).includes(
+        return this.allSurahList.filter((item) => {
+          const predicateTranslation = __normalizeText(item.translation).includes(
             __normalizeText(this.searchText)
           )
-          let predicateLatin = __normalizeText(item.latin).includes(
+          const predicateLatin = __normalizeText(item.latin).includes(
             __normalizeText(this.searchText)
           )
 
@@ -73,25 +73,16 @@ export default {
       } else return this.allSurahList
     }
   },
-  mounted () {
-    this.setHeaderTitle('Daftar Surat')
-    this.fetchSurahInfo()
-  },
-  methods: {
-    ...mapMutations([
-      'setHeaderTitle'
-    ]),
-    ...mapActions([
-      'fetchAllSurah'
-    ]),
-    fetchSurahInfo () {
-      this.fetchAllSurah({
-        success: this.onSuccess
+  async asyncData() {
+    const data = await getAllSurah()
+    return {
+      allSurahList: data.data.surah_info.map((item, idx) => {
+        return Object.assign({}, item, { index: idx + 1 })
       })
-    },
-    onSuccess () {
-      this.loading = false
     }
+  },
+  fetch({ store }) {
+    store.commit('setHeaderTitle', 'Daftar Surat')
   }
 }
 </script>
