@@ -1,5 +1,12 @@
 /* eslint-env jest */
-import { getAllSurah, getSurahById, getAyatKursi, getAsmaulHusna, getDailyDoa } from '~/services/index'
+import {
+  getAllSurah,
+  getSurahById,
+  getAyatKursi,
+  getAsmaulHusna,
+  getDailyDoa
+} from '~/services/index'
+import axios from 'axios'
 
 beforeEach(() => {
   jest.resetAllMocks()
@@ -14,9 +21,7 @@ describe('service index', () => {
     const expectedResponse = {
       data: 'dummy'
     }
-    global.fetch = jest.fn().mockResolvedValueOnce({
-      json: () => expectedResponse
-    })
+    axios.get.mockResolvedValueOnce(expectedResponse)
     const response = await getAllSurah()
     expect(response).toEqual(expectedResponse)
     done()
@@ -26,9 +31,7 @@ describe('service index', () => {
     const expectedResponse = {
       data: 'dummy'
     }
-    global.fetch = jest.fn().mockResolvedValueOnce({
-      json: () => expectedResponse
-    })
+    axios.get.mockResolvedValueOnce(expectedResponse)
     const response = await getSurahById('some-id')
     expect(response).toEqual(expectedResponse)
     done()
@@ -38,9 +41,7 @@ describe('service index', () => {
     const expectedResponse = {
       data: 'dummy'
     }
-    global.fetch = jest.fn().mockResolvedValueOnce({
-      json: () => expectedResponse
-    })
+    axios.get.mockResolvedValueOnce(expectedResponse)
     const response = await getAyatKursi()
     expect(response).toEqual(expectedResponse)
     done()
@@ -50,9 +51,7 @@ describe('service index', () => {
     const expectedResponse = {
       data: 'dummy'
     }
-    global.fetch = jest.fn().mockResolvedValueOnce({
-      json: () => expectedResponse
-    })
+    axios.get.mockResolvedValueOnce(expectedResponse)
     const response = await getAsmaulHusna()
     expect(response).toEqual(expectedResponse)
     done()
@@ -62,9 +61,7 @@ describe('service index', () => {
     const expectedResponse = {
       data: 'dummy'
     }
-    global.fetch = jest.fn().mockResolvedValueOnce({
-      json: () => expectedResponse
-    })
+    axios.get.mockResolvedValueOnce(expectedResponse)
     const response = await getDailyDoa()
     expect(response).toEqual(expectedResponse)
     done()
